PRBS Generator and Detector
Known good working PRBS generator and detector.
tb_prbs_gen_det.vhd  has a PRBS5 and PRBS16 generator.  I've tested both and they work.   
prbs_det.vhd has a weird quirk when it first starts, if the valid_data flag isn't in line with the reset it will get a bunch of errors, then once it restarts and trains the LFSR on the incoming bits it will gain sync no problem.  Not sure why the first set of 100 errors is happening.