Basys3 PRBS Gen Det
This is the implementation of the PRBS Generator and Detector on the Basys3 board.  This is working with a PN5 sequence.  It did not work with a PRBS16 bit sequence for some reason.   Since I am only worried about the 32 bit PRBS sequence at this point (assuming a 64 bit Viterbi word) this is the working version.   I will work on debugging why PN16 sequence doesn't work later.   
This Revision has a reset, generate error buttons and generate data switch, that will display errors on the 7 segment display.  PRBS input and outputs are on JB1,JB7 and JC1 JC7 pmod connectors.  1->1 PRBS bit   7->7 PRBS valid data