PRBSGen->FdFwdViterbiEncoder->FdFwdViterbiDec
This is a sort of working test bench.  There is an issue where a word is decoded correctly other than 1 or 2 errors at the beginning and/or end of the word.  Other than that it does the correct decoding.  It also stops decoding after about 7-8 words.  I can't figure out why.   Using this as a point of reference to know where I left off before I make some changes.