Failed hardware implementation
The basys3_viterbi_prbs_gen_det.vhd is so close to working on hardware, but in the simulation you can see where it kind of goes wrong.  I just can't get the timing right to make it pass.  I'm going to need some more help to try and figure this out.  i'm running out of ideas how to  fix what is happening.