Debug Start
I am having a hell of a time figuring out what the hell is keeping this from working!   I have a simulation tb_basys3_viterbi_prbs_gen_det that will sort of work, it should be outputting 0x4B3E3750, but it's outputting 0x0B3E3750!! Talk about swearing like a sailor!   I can't figure out how to get this to not miss the last bit, in other words the 1 just before starting another 31 bits.  it goes 0010110011111000110111010100001, i'm missing the very last 1 and it's causing it to output the wrong thing at the decoder.