Working Simulation with FIFO Buffers
This is a working simulation of FIFO buffers in between all data output modules.  This worked only after tweeking the viterbi encoder to sent a enc_not_ready signal when it his the 32nd bit.  This is the bit that needs to terminate the word.  It seems to be working fine, but I need to put this in hardware and see what happens.   this simulation gets sync after just over 15ms.  This due to the clock now being about ~6MHz